refactor: remove 'models' folder from 'src'
