+++ .gitignore
@@ -10,7 +10,7 @@ pnpm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
-!/src/models/logs
+!/src/logs
 
 # OS
 .DS_Store
