+++ src/app.module.ts
@@ -1,18 +1,18 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 
-import { LogConfig } from './models/logs/log.config';
-import { LogsController } from './models/logs/controllers/logs.controller';
-import { LogsService } from './models/logs/services/logs.service';
-import { LogRepository } from './models/logs/log.repository';
-import { LogSchema } from './models/logs/schemas/log.schema';
+import { LogConfig } from './logs/log.config';
+import { LogsController } from './logs/controllers/logs.controller';
+import { LogsService } from './logs/services/logs.service';
+import { LogRepository } from './logs/log.repository';
+import { LogSchema } from './logs/schemas/log.schema';
 import { CommonConfig } from './common/config';
-import { MachineConfig } from './models/machines/machine.config';
-import { MachineSchema } from './models/machines/schemas/machine.schema';
-import { MachinesController } from './models/machines/controllers/machines.controller';
-import { MachinesService } from './models/machines/services/machines.service';
-import { MachineRepository } from './models/machines/machine.repository';
-import { MachinesCommonService } from './models/machines/common/services/machines.service';
+import { MachineConfig } from './machines/machine.config';
+import { MachineSchema } from './machines/schemas/machine.schema';
+import { MachinesController } from './machines/controllers/machines.controller';
+import { MachinesService } from './machines/services/machines.service';
+import { MachineRepository } from './machines/machine.repository';
+import { MachinesCommonService } from './machines/common/services/machines.service';
 
 @Module({
   imports: [
