+++ src/logs/log.repository.ts
@@ -3,7 +3,7 @@ import { InjectModel } from '@nestjs/mongoose';
 
 import { Model } from 'mongoose';
 
-import { LogDTO } from 'src/models/logs/dto/logs.dto';
+import { LogDTO } from 'src/logs/dto/logs.dto';
 import { Log } from './interfaces/log.interface';
 import { LogConfig } from './log.config';
