+++ src/machines/dto/machines.dto.ts
