+++ src/machines/interfaces/machine.interface.ts
