+++ src/machines/machine.config.ts
