+++ src/machines/schemas/machine.schema.ts
